Extract BrachInjector from key transformer
