+++ tests/Unit/Service/Yaml/Transformer/KeysTransformerTest.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Aeliot\Bundle\TransMaintain\Test\Unit\Service\Yaml\Transformer;
 
+use Aeliot\Bundle\TransMaintain\Service\Yaml\BranchInjector;
 use Aeliot\Bundle\TransMaintain\Service\Yaml\Transformer\KeysTransformer;
 use Generator;
 use PHPUnit\Framework\TestCase;
@@ -20,7 +21,7 @@ final class KeysTransformerTest extends TestCase
      */
     public function testTransform(array $expected, array $income): void
     {
-        self::assertEquals($expected, (new KeysTransformer())->transform($income));
+        self::assertEquals($expected, (new KeysTransformer(new BranchInjector()))->transform($income));
     }
 
     public function getDataForTestTransform(): Generator
